% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortExplorer.R
\docType{package}
\name{CohortExplorer-package}
\alias{CohortExplorer}
\alias{CohortExplorer-package}
\title{CohortExplorer: Explorer of Profiles of Patients in a Cohort}
\description{
This software tool is designed to extract data from a randomized subset of individuals within a cohort and make it available for exploration in a shiny application environment. It retrieves date-stamped, event-level records from one or more data sources that represent patient data in the Observational Medical Outcomes Partnership (OMOP) data model format. This tool features a user-friendly interface that enables users to efficiently explore the extracted profiles, thereby facilitating applications, such as reviewing structured profiles.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/CohortExplorer/}
  \item \url{https://github.com/OHDSI/CohortExplorer}
  \item Report bugs at \url{https://github.com/OHDSI/CohortExplorer/issues}
}

}
\author{
\strong{Maintainer}: Gowtham Rao \email{rao@ohdsi.org}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
