% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportCohortExplorerAppFiles.R
\name{exportCohortExplorerAppFiles}
\alias{exportCohortExplorerAppFiles}
\title{Copy shiny app files}
\usage{
exportCohortExplorerAppFiles(exportFolder)
}
\arguments{
\item{exportFolder}{The folder where the output will be exported to. If this folder
does not exist it will be created.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{                      Nothing is returned to the environment. The function creates the
                               required files for the shiny application in the export folder.
}\if{html}{\out{</div>}}
}
\description{
Copy shiny app files.
}
\examples{
\dontrun{
exportCohortExplorerAppFiles(
  exportFolder = "output"
)
}

}
