% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractPersonLevelData.R
\name{extractaPersonLevelData}
\alias{extractaPersonLevelData}
\title{Extract person level data for cohort}
\usage{
extractaPersonLevelData(
  connectionDetails,
  cohortDatabaseSchema = "cohort",
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  tempEmulationSchema = NULL,
  cohortTable = "cohort",
  cohortDefinitionId,
  sampleSize = 25,
  personIds = NULL
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{cohortDatabaseSchema}{Schema name where your cohort tables reside. Note that for SQL Server,
this should include both the database and schema name, for example
'scratch.dbo'.}

\item{cdmDatabaseSchema}{Schema name where your omop cdm tables with person level data reside.
Note that for SQL Server, this should include both the database and schema name, for example
'scratch.dbo'.}

\item{vocabularyDatabaseSchema}{Schema name where your omop vocabulary tables reside. It maybe the cdmDatabaseSchema.
Note that for SQL Server, this should include both the database and schema name, for example
'scratch.dbo'.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support
temp tables. To emulate temp tables, provide a schema with write
privileges where temp tables can be created.}

\item{cohortTable}{The name of the cohort table.}

\item{cohortDefinitionId}{The cohort id to extract records.}

\item{sampleSize}{(Optional, default = 20) The number of persons to randomly sample. Ignored, if personId is given.}

\item{personIds}{(Optional) An array of personId's to look for in Cohort table and CDM.}
}
\description{
Given an instantiated cohort table, extract all person level data from omop cdm tables
for persons found in the cohort.
}
\examples{
\dontrun{
connectionDetails <- createConnectionDetails(
  dbms = "postgresql",
  server = "ohdsi.com",
  port = 5432,
  user = "me",
  password = "secure"
)

extractaPersonLevelData(
  connectionDetails = connectionDetails,
  cohortDefinitionId = 1234
)
}

}
