% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortExplorer-package.R, R/CohortExplorer.R
\docType{package}
\name{CohortExplorer-package}
\alias{CohortExplorer}
\alias{CohortExplorer-package}
\title{CohortExplorer: An R package with a Shiny viewer to explore profiles of patients in a cohort.}
\description{
An R package with a Shiny viewer to explore profiles of patients in a cohort.

An R package with a Shiny viewer to explore profiles of patients in a cohort.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/CohortExplorer/}
  \item \url{https://github.com/OHDSI/CohortExplorer}
  \item Report bugs at \url{https://github.com/OHDSI/CohortExplorer/issues}
}


Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/CohortExplorer/}
  \item \url{https://github.com/OHDSI/CohortExplorer}
  \item Report bugs at \url{https://github.com/OHDSI/CohortExplorer/issues}
}

}
\author{
\strong{Maintainer}: Gowtham Rao \email{rao@ohdsi.org}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
