% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportCohortExplorerAppFiles.R
\name{exportCohortExplorerAppFiles}
\alias{exportCohortExplorerAppFiles}
\title{Copy shiny app files}
\usage{
exportCohortExplorerAppFiles(exportFolder)
}
\arguments{
\item{exportFolder}{The folder where the output will be exported to. If this folder
does not exist it will be created.}
}
\description{
Copy shiny app files.
}
\examples{
\dontrun{
exportCohortExplorerAppFiles(
  exportFolder = "output"
)
}

}
